# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/user/new-password": {
      "post": {
        "operationId": "UserController_updateNewPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tạo mật khẩu mới thành công"
          }
        },
        "summary": "Tạo mật khẩu mới",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/change-password": {
      "post": {
        "operationId": "UserController_changePassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email của người dùng"
                  },
                  "oldPassword": {
                    "type": "string",
                    "description": "Mật khẩu cũ"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "Mật khẩu mới"
                  }
                },
                "required": [
                  "email",
                  "oldPassword",
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đổi mật khẩu thành công"
          },
          "401": {
            "description": "Mật khẩu cũ không đúng"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đổi mật khẩu",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/profile": {
      "get": {
        "operationId": "UserController_getCurrentUserProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lấy thông tin profile thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin profile của người dùng hiện tại",
        "tags": [
          "User"
        ]
      },
      "put": {
        "operationId": "UserController_updateCurrentUserProfile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật profile thành công"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật thông tin profile của người dùng hiện tại",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/{userId}/profile": {
      "get": {
        "operationId": "UserController_getUserProfile_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID của người dùng",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin profile thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin profile của người dùng theo ID",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/avatar-image": {
      "post": {
        "operationId": "UserController_uploadAvatarImage_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Avatar image upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload avatar image successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "asset_id": "some-asset-id",
                      "public_id": "some-public-id",
                      "url": "https://res.cloudinary.com/..."
                    },
                    "message": "Tải lên tệp tin thành công"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload avatar image for current user",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/cover-image": {
      "post": {
        "operationId": "UserController_uploadCoverImage_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Cover image upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload cover image successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "asset_id": "some-asset-id",
                      "public_id": "some-public-id",
                      "url": "https://res.cloudinary.com/..."
                    },
                    "message": "Tải lên tệp tin thành công"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload cover image for current user",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/v1/user/skills": {
      "put": {
        "operationId": "UserController_updateSkillLevels_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Skill levels",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "skillLevels": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "enum": [
                        "beginner",
                        "intermediate",
                        "advanced",
                        "professional"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật skill levels thành công"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật trình độ kỹ năng cho từng môn thể thao",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/favorite-sport/{sportType}": {
      "delete": {
        "operationId": "UserController_removeFavoriteSport_v1",
        "parameters": [
          {
            "name": "sportType",
            "required": true,
            "in": "path",
            "description": "Tên môn thể thao cần xóa",
            "schema": {
              "enum": [
                "football",
                "basketball",
                "tennis",
                "badminton",
                "volleyball",
                "table_tennis",
                "swimming",
                "running",
                "cycling",
                "yoga",
                "gym",
                "martial_arts",
                "boxing",
                "golf",
                "rock_climbing"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa môn thể thao yêu thích thành công"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa 1 môn thể thao khỏi danh sách yêu thích và skillLevels",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/friends/search": {
      "get": {
        "operationId": "UserController_getFriendsByKey_v1",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách bạn bè phù hợp",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FriendSimpleDto"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tìm kiếm bạn bè theo tên (key trong fullname, không phân biệt hoa thường, có phân trang)",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/{userId}/basic-info": {
      "get": {
        "operationId": "UserController_getUserBasicInfo_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID của người dùng",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin cơ bản thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBasicInfoDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin cơ bản của người dùng theo ID",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/fcm-token": {
      "put": {
        "operationId": "UserController_updateFcmToken_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fcmToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật FCM token thành công"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật FCM token cho user",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/device-language": {
      "put": {
        "operationId": "UserController_updateDeviceLanguage_v1",
        "parameters": [
          {
            "name": "lang",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cập nhật ngôn ngữ thiết bị thành công"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc thiếu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật ngôn ngữ thiết bị cho user",
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "Swagger docs for the project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Map": {
        "type": "object",
        "properties": {}
      },
      "ResponseProfileDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID của người dùng"
          },
          "email": {
            "type": "string",
            "description": "Email của người dùng"
          },
          "firstName": {
            "type": "string",
            "description": "Tên"
          },
          "lastName": {
            "type": "string",
            "description": "Họ"
          },
          "fullName": {
            "type": "string",
            "description": "Tên đầy đủ"
          },
          "avatar": {
            "type": "string",
            "description": "Ảnh đại diện"
          },
          "coverImage": {
            "type": "string",
            "description": "Ảnh bìa"
          },
          "bio": {
            "type": "string",
            "description": "Tiểu sử"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string",
            "description": "Ngày sinh"
          },
          "phone": {
            "type": "string",
            "description": "Số điện thoại"
          },
          "location": {
            "type": "object",
            "description": "Địa chỉ"
          },
          "favoritesSports": {
            "description": "Các môn thể thao yêu thích",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "skillLevels": {
            "description": "Mức độ kỹ năng",
            "allOf": [
              {
                "$ref": "#/components/schemas/Map"
              }
            ]
          },
          "friendsCount": {
            "type": "number",
            "description": "Số lượng bạn bè"
          },
          "followingCount": {
            "type": "number",
            "description": "Số lượng đang theo dõi"
          },
          "followersCount": {
            "type": "number",
            "description": "Số lượng người theo dõi"
          },
          "joinedGroupsCount": {
            "type": "number",
            "description": "Số lượng nhóm đã tham gia"
          },
          "isActive": {
            "type": "boolean",
            "description": "Trạng thái hoạt động"
          },
          "isVerified": {
            "type": "boolean",
            "description": "Trạng thái xác thực"
          },
          "roles": {
            "description": "Vai trò",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Ngày tạo"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Ngày cập nhật"
          }
        },
        "required": [
          "_id",
          "email",
          "firstName",
          "lastName",
          "fullName",
          "friendsCount",
          "followingCount",
          "followersCount",
          "joinedGroupsCount",
          "isActive",
          "isVerified",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Tên"
          },
          "lastName": {
            "type": "string",
            "description": "Họ"
          },
          "avatar": {
            "type": "string",
            "description": "Ảnh đại diện"
          },
          "coverImage": {
            "type": "string",
            "description": "Ảnh bìa"
          },
          "bio": {
            "type": "string",
            "description": "Tiểu sử"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string",
            "description": "Ngày sinh"
          },
          "phone": {
            "type": "string",
            "description": "Số điện thoại"
          },
          "location": {
            "description": "Địa chỉ",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocationDto"
              }
            ]
          },
          "favoritesSports": {
            "description": "Các môn thể thao yêu thích",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "skillLevels": {
            "description": "Mức độ kỹ năng",
            "allOf": [
              {
                "$ref": "#/components/schemas/Map"
              }
            ]
          }
        }
      },
      "UserBasicInfoDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID của người dùng",
            "example": "685a2443342f516009019c71"
          },
          "fullName": {
            "type": "string",
            "description": "Họ và tên đầy đủ",
            "example": "John Doe"
          },
          "avatar": {
            "type": "object",
            "description": "URL avatar của người dùng",
            "example": "https://example.com/avatar.jpg",
            "nullable": true
          }
        },
        "required": [
          "userId",
          "fullName",
          "avatar"
        ]
      },
...
    }
  }
}